GH-418 Extracted the modal window on the Details page into a separate component
