+++ front-end/src/pages/Details/DetailsFirstSection/styles.module.css
@@ -24,43 +24,6 @@
     width: 16px;
 }
 
-.List {
-    margin-block: 30px;
-    overflow: hidden;
-}
-
-.Collapse {
-    border: none;
-    border-bottom: 1px solid #0505050f;
-    border-radius: 0;
-    background-color: #ffffff;
-}
-
-.Collapse :global(.ant-collapse-header):hover {
-    cursor: default;
-}
-
-.HeapDumpAccordionHeader {
-    display: flex;
-    justify-content: space-between;
-    padding-inline: 8px 16px;
-}
-
-.Component {
-    cursor: text;
-}
-
-.HeapDumpAccordionBody {
-    display: grid;
-    grid-template-columns: max-content minmax(0, 1fr);
-    gap: 20px;
-    padding-inline: 28px 16px;
-}
-
-.LoaderWrapper{
-    margin-block: 40px;
-}
-
 @media all and (max-width: 1220px) {
     .MainTitle {
         text-align: left;
